Sequence protocol
A list is internally represented by a nested two-element quote, where
first is a list element, and the second is a quote for the rest (tail)
of the list.

A 'false quote in the tail element denotes the end of the list.

The list doesn't have to be finite. It doesn't have to know its length.
Just enough to iterate through it.