Change the semantics of sequences to a quote
Previously, a sequence was expected to be a tuple of head and tail.
It limited and complicated multiple-value sequences, and had a bug when
a dynamic sequence couldn't be described as a single-word quote just
like e.g. a monotonic sequence that doesn't need an explicit input.
That also simplified the notation for other sequence generators.